users.js includes helper functions to abstract functionality from api endpoints
